minor: get rid of unnecessary restful server lock
